refactor(frontend): restructure directories in src
