--- frontend/src/coordinate/CoordinatePage.tsx
@@ -1,7 +1,0 @@
-import React from 'react';
-
-// URL: /coordinate/
-
-export const CoordinatePage: React.FC = () => {
-  return <div />;
-};
