+++ frontend/src/functions/AccountInfoSubmitValidation.ts
