+++ frontend/src/functions/GetCookieArray.ts
