+++ frontend/src/functions/HttpRequest.ts
