+++ frontend/src/pages/ContestPage.css
