+++ frontend/src/pages/NotFoundPage.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-export const NotFound: React.FC = () => {
+export const NotFoundPage: React.FC = () => {
   return (
     <div>
       <p>URLが存在しません</p>
